Makefile.PL: remove install overwrite checks
This seems to be checking if an existing version exists on the system
and, if so, require that the install overwrites it, and refuse anything
else.

That breaks if you're dh-make-perl installing to a private dir to make
an installable package. None of this seems to be needed in that case.